updated the routes
